+++ src/lib/posix/read.c
@@ -3,7 +3,7 @@
 PUBLIC int read(
 int fd,
 char *buffer,
-unsigned nbytes
+size_t nbytes
 ){
   return(callm1(FS, READ, fd, nbytes, 0, buffer, NIL_PTR, NIL_PTR));
 }
